requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
numpy==2.4.2
opencv-python==4.13.0.92
pyserial==3.5
requests==2.32.5
urllib3==2.6.3
python-dotenv